+++ tests/test_mmi_format.py
@@ -154,7 +154,8 @@ def test_comma():
     result = next(extract_mmi_line(lst))
     assert result == {'docid': '0000', 'filename': '0000.tx', 'matchedtext': 'Pain, chest',
                       'conceptstring': 'Chest Pain', 'cui': 'C0008031', 'preferredname': 'Pain, Chest', 'start': 0,
-                      'length': 11, 'end': 11, 'evid': None, 'negated': 0, 'pos': '', 'semantictype': 'sosy', 'sosy': 1}
+                      'length': 11, 'end': 11, 'evid': None, 'negated': 0, 'pos': '', 'semantictype': 'sosy',
+                      'all_semantictypes': 'sosy', 'all_sources': '', 'sosy': 1}
 
 
 @pytest.mark.parametrize('triggerinfo_text, concept, loc, locpos, text, pos, neg', [
@@ -188,17 +189,6 @@ def test_triggerinfo(triggerinfo_text, concept, loc, locpos, text, pos, neg):
 def test_empty_mmi_line(mmi_line):
     with pytest.raises(StopIteration):
         next(extract_mmi_line(mmi_line))
-
-
-@pytest.mark.parametrize('line, exp_logtext', [
-    ('0000.tx|MMI|2.30|Chest Pain|C0008031|[sosy]'
-     '|"Pain, Chest"-text-0-"Pain, chest"--0|text|11|C23.888.592.612.233',
-     'Unknown format of length 1 for positional info'),
-])
-def test_has_invalid_length(caplog, line, exp_logtext):
-    with caplog.at_level(logging.ERROR):
-        _ = list(extract_mmi_line(line.split('|')))
-    assert exp_logtext in caplog.text
 
 
 @pytest.mark.parametrize('text, exp_length, exp_triggerinfo', [